Refactor aggregate with count.
