+++ tiledb/sm/query/readers/aggregators/sum_aggregator.h
@@ -33,9 +33,10 @@
 #ifndef TILEDB_SUM_AGGREGATOR_H
 #define TILEDB_SUM_AGGREGATOR_H
 
-#include "tiledb/sm/query/readers/aggregators/aggregate_sum.h"
+#include "tiledb/sm/query/readers/aggregators/aggregate_with_count.h"
 #include "tiledb/sm/query/readers/aggregators/field_info.h"
 #include "tiledb/sm/query/readers/aggregators/iaggregator.h"
+#include "tiledb/sm/query/readers/aggregators/sum_type.h"
 
 namespace tiledb {
 namespace sm {
@@ -115,8 +116,8 @@ class SumAggregator : public OutputBufferValidator, public IAggregator {
   /** Field information. */
   const FieldInfo field_info_;
 
-  /** AggregateSum to do summation of AggregateBuffer data. */
-  AggregateSum<T> summator_;
+  /** AggregateWithCount to do summation of AggregateBuffer data. */
+  AggregateWithCount<T> summator_;
 
   /** Computed sum. */
   std::atomic<typename sum_type_data<T>::sum_type> sum_;
